backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# WebSocket
python-socketio==5.10.0
aiohttp==3.9.1

# Data Validation
pydantic==2.5.2
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0

# Scientific Computing
numpy==1.26.2
gymnasium==0.29.1

# XML Processing
lxml==4.9.3

# Logging
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Code Quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1
